+++ CHANGELOG.md
@@ -154,3 +154,4 @@
 - `NBTString`
 - Casting
 - Attribute setting / getting
+- `start` script tag